added 'processed data' management to manage only the data processed for each data subject

+++ PRIAM-Data-service/src/main/java/priam/data/priamdataservice/entities/consent/ProcessedDataKey.java
@@ -3,6 +3,6 @@ package priam.data.priamdataservice.entities.consent;
 import java.io.Serializable;
 
 public class ProcessedDataKey implements Serializable {
-    private Long dataSubjectId;
-    private Long dataId;
+    private int dataSubjectId;
+    private int dataId;
 }
